FDG-6282 refactoring section
