+++ src/layouts/explainer/sections/national-debt/tracking-the-debt/tracking-the-debt.jsx
@@ -1,0 +1,33 @@
+import React from 'react';
+import CustomLink from '../../../../../components/links/custom-link/custom-link';
+import { analyticsClickHandler } from '../../../explainer-helpers/national-debt-helper';
+
+
+const TrackingTheDebt = () => {
+  const fiscalService = (
+    <CustomLink
+      url={"https://www.fiscal.treasury.gov/"}
+      onClick={() =>
+        analyticsClickHandler("Citation Click", "Tracking the Debt")
+      }
+    >
+      Bureau of the Fiscal Service
+    </CustomLink>
+  );
+
+  return (
+    <>
+      <p>
+        Created in 2012 and operating under the Department of the Treasury, the{" "}
+        {fiscalService} manages all federal payments and collections and
+        provides government-wide accounting and reporting services. A primary
+        function of the Fiscal Service is to account for and report the national
+        debt, as dictated by the U.S. Constitution, which states that “regular
+        Statement and Account of the Receipts and Expenditures of all public
+        Money shall be published from time to time.”
+      </p>
+    </>
+  );
+}
+
+export default TrackingTheDebt;
